+++ SourceCode/Services/ListboxItem.cs
@@ -1,3 +1,0 @@
-﻿namespace ModulesRegistry.Services;
-
-public record ListboxItem(int Id, string Description) { public bool IsHidden { get; set; } }